I completed the swapPhoto function.
We get the img src of the photo and replace it from what was left on it before.

We set the text of the picture's information to be shown to the user.

If the currentIndex is less than the array's length - 1, then we add 1 to the counting number. If we reach th last picture, then the picture gets swapped to the zeroth index of the array.